making the default "rendering" more boring
Instead of an action, where we send up old data AND new data... then
merge them on the server, we now merge the data in JavaScript. Then,
rendering the component (when there is no action) is a boring
"Hey, here is some data you should render the element with".

This means the data can be arbitrarily changed, which is fine. The
key this is that the "model update" (where we just re-render the
component with a new set of data) will always be a "safe" operation
where no changes are made on the server. An action should be used
for those.